% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_data.R
\name{append.cnv}
\alias{append.cnv}
\title{append CNV from file name vector}
\usage{
append.cnv(x)
}
\arguments{
\item{x}{file name vector}
}
\value{
cnv table with Entres_Gene_Id
}
\description{
append CNV from file name vector
}
\examples{

}
