% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.R
\name{append_all_maf}
\alias{append_all_maf}
\title{Append all_maf file}
\usage{
append_all_maf(new_maf_file, old_maf_file)
}
\arguments{
\item{new_maf_file}{file path to new patient data file}

\item{old_maf_file}{file path to old patient data file}
}
\value{
appended all_maf file in /home/molpath/cbioportal_data/study-dir/
}
\description{
Append all_maf file
}
\examples{
\dontrun{
append_all_maf()
}
}
