% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.R
\name{append_discrete_cnv}
\alias{append_discrete_cnv}
\title{Append discrete cnv file}
\usage{
append_discrete_cnv(new_discrete_cnv_table, old_discrete_cnv_file)
}
\arguments{
\item{new_discrete_cnv_table}{new discrete cnv data table generated by \code{\link{discretize_cnv}}}

\item{old_discrete_cnv_file}{file path to old discrete cnv file}
}
\value{
appended case list data files overriding the original data files
}
\description{
Append discrete cnv file
}
\examples{
\dontrun{
append_discrete_cnv(new_discrete_cnv_table, old_discrete_cnv_file)
}

}
