% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.R
\name{append_fusion}
\alias{append_fusion}
\title{Append fusion file}
\usage{
append_fusion(new_fusion_df, old_fusion_file)
}
\arguments{
\item{new_fusion_df}{new fusion data table generated by \code{\link{make.fusion.df}}}

\item{old_fusion_file}{file path to old fusion file}
}
\value{
appended fusion data files overriding the original data files
}
\description{
Append fusion file
}
\examples{
\dontrun{
append_fusion(new_fusion_df, old_fusion_file)
}

}
