% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.R
\name{append_list}
\alias{append_list}
\title{Append case list file}
\usage{
append_list(new_list_directory, old_list_directory)
}
\arguments{
\item{new_list_directory}{file path to new case list directory}

\item{old_list_directory}{file path to old case list directory}
}
\value{
appended case list data files overriding the original data files
}
\description{
Append case list file
}
\examples{
\dontrun{
append_list(new_list_directory, old_list_directory)
}
}
