% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.R
\name{append_patient}
\alias{append_patient}
\title{Append patient data file}
\usage{
append_patient(new_patient_file, old_patient_file)
}
\arguments{
\item{new_patient_file}{file path to new patient data file}

\item{old_patient_file}{file path to old patient data file}
}
\value{
appended patient data file in /home/molpath/cbioportal_data/study-dir/
}
\description{
Append patient data file
}
\examples{
\dontrun{
append_all_maf(new_patient_file, old_patient_file)
}
}
