% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.R
\name{append_sample}
\alias{append_sample}
\title{Append sample data file}
\usage{
append_sample(new_sample_file, old_sample_file)
}
\arguments{
\item{new_sample_file}{file path to new patient data file}

\item{old_sample_file}{file path to old patient data file}
}
\value{
appended patient data file in /home/molpath/cbioportal_data/study-dir/
}
\description{
Append sample data file
}
\examples{
\dontrun{
append_all_maf(new_sample_file, old_sampe_file)
}
}
