% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_data.R
\name{cnv_table}
\alias{cnv_table}
\title{Append cnv data to data.frame cnv_df}
\usage{
cnv_table(x)
}
\arguments{
\item{x}{file name}
}
\value{
cnv data.frame
}
\description{
Append cnv data to data.frame cnv_df
}
