% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/renameFile.R
\name{copyRenameVcf}
\alias{copyRenameVcf}
\title{copy vcf to rename vcf into output directory}
\usage{
copyRenameVcf(inputVcfDirectory, outputVcfDirectory)
}
\arguments{
\item{inputVcfDirectory}{string of path to input directory}

\item{outputVcfDirectory}{string of path to output directory}
}
\value{
copy vcf to rename vcf into output directory
}
\description{
copy vcf to rename vcf into output directory
}
\examples{
\dontrun{
copyRenameVcf(vcf, "vcfRename")
}


}
