% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_extract.R
\name{extract_ion2cbio}
\alias{extract_ion2cbio}
\title{extract vcf filtered and tsv filtered files from ion torrent zip files}
\usage{
extract_ion2cbio(input_directory, output_directory)
}
\arguments{
\item{input_directory}{string of path to input directory}

\item{output_directory}{string of path to output directory}
}
\value{
extract vcf and tsv files to vcf and tsv directory
}
\description{
extract vcf filtered and tsv filtered files from ion torrent zip files
}
\examples{
\dontrun{
extract_ion2cbio("/home/molpath/cbioportal_data/zip", "/home/molpath/cbioportal_data/output")
}

}
