% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion.R
\name{find.Entres_Gene_Id}
\alias{find.Entres_Gene_Id}
\title{make translocation data}
\usage{
find.Entres_Gene_Id(gene_name)
}
\arguments{
\item{gene_name}{gene name of OCAv3 fusion}
}
\value{
Entres_Gene_Id
}
\description{
make translocation data
}
\examples{
\dontrun{
find.Entres_Gene_Id("ALK")
}
}
