% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion.R
\name{make.fusion.data.file}
\alias{make.fusion.data.file}
\title{make translocation data}
\usage{
make.fusion.data.file(fusion_df, output_file)
}
\arguments{
\item{fusion_df}{tibble with fusion data}

\item{output_file}{path/to/output/file}
}
\value{
fusion data.file
}
\description{
make translocation data
}
\examples{
\dontrun{
make.fusion.data.file(fusion_df, output_file)
}
}
