% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion.R
\name{make.fusion.df}
\alias{make.fusion.df}
\title{make translocation data}
\usage{
make.fusion.df(list_tsv_file)
}
\arguments{
\item{list_tsv_file}{list of tsv file name}
}
\value{
tibble with fusion data
}
\description{
make translocation data
}
\examples{
\dontrun{
make.fusion.df(list_tsv_file)
}
}
