% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_list.R
\name{make_case_list}
\alias{make_case_list}
\title{Make case list files}
\usage{
make_case_list(x)
}
\arguments{
\item{x}{path to maf directory}
}
\value{
case list files (case_list_sequenced, cnvseq, cna) in study-dir/case_lists directory
}
\description{
Make case list files
}
\examples{
\dontrun{
make_case_list(x)
}
}
