% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampleData.R
\name{make_sample_data_file}
\alias{make_sample_data_file}
\title{Make sample data files}
\usage{
make_sample_data_file(pathInput, pathOutput)
}
\arguments{
\item{pathInput}{path to maf directory}

\item{pathOutput}{path to output}
}
\value{
sample data files
}
\description{
Make sample data files
}
\examples{
\dontrun{
make_sample_data_file("input/directory", "output/directory" )
}
}
