% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_data.R
\name{merge_cnv}
\alias{merge_cnv}
\title{merge CNV from tsv files}
\usage{
merge_cnv(f_list)
}
\arguments{
\item{x}{character vector of tsv file name use list.files}
}
\value{
cnv table with Entres_Gene_Id
}
\description{
merge CNV from tsv files
}
\examples{
f_list <- list.files('/home/molpath/cbioportal_data/output/tsv/', full.names = TRUE)
merge_cnv(f_list)
}
