% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fusion.R
\name{tidy_fusion}
\alias{tidy_fusion}
\title{make translocation data}
\usage{
tidy_fusion(tsv_file)
}
\arguments{
\item{tsv_file}{path/to/tsv/file}
}
\value{
tidy fusion data.frame
}
\description{
make translocation data
}
\examples{
\dontrun{
tidy_fusion(tsv_file)
}
}
