% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_data.R
\name{tsv2table}
\alias{tsv2table}
\title{Append cnv data to data.frame cnv_df}
\usage{
tsv2table(x)
}
\arguments{
\item{x}{file name}
}
\value{
cnv data.frame
}
\description{
Append cnv data to data.frame cnv_df
}
\examples{
\dontrun{
tsv2table("M19-0000")
}
}
