% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{all_maf}
\alias{all_maf}
\title{merge maf from vcf file}
\usage{
all_maf(input_directory)
}
\arguments{
\item{input_directory}{character of directory of vcf files}
}
\value{
merged maf file
}
\description{
merge maf from vcf file
}
\examples{


}
