% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.R
\name{append_all_maf}
\alias{append_all_maf}
\title{Append all_maf file}
\usage{
append_all_maf(pathTodirectory, new_maf_file, old_maf_file)
}
\arguments{
\item{pathTodirectory}{path to directory of new maf}

\item{new_maf_file}{file path to new maf file}

\item{old_maf_file}{file path to old maf file}
}
\value{
appended all_maf file overriding the original data file
}
\description{
Append all_maf file
}
\examples{
\dontrun{
append_all_maf(pathTodirectory, new_maf_file, old_maf_file)
}
}
