% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.R
\name{append_list}
\alias{append_list}
\title{Append case list file}
\usage{
append_list(new_list_directory, old_list_directory)
}
\arguments{
\item{new_list_directory}{file path to new patient data file}

\item{old_list_directory}{file path to old patient data file}
}
\value{
appended patient data file in /home/molpath/cbioportal_data/study-dir/
}
\description{
Append case list file
}
\examples{
\dontrun{
append_all_maf(new_sample_file, old_sampe_file)
}
}
