% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append.R
\name{append_sample}
\alias{append_sample}
\title{Append sample data file}
\usage{
append_sample(pathTodirectory, new_sample_file, old_sample_file)
}
\arguments{
\item{pathTodirectory}{directory path to data file}

\item{new_sample_file}{file path to new sample data file}

\item{old_sample_file}{file path to old sample data file}
}
\value{
appended sample data file overriding the original data file
}
\description{
Append sample data file
}
\examples{
\dontrun{
append_sample(pathTodirectory, new_sample_file, old_sampe_file)
}
}
