% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cnv_data.R
\name{discretize_cnv}
\alias{discretize_cnv}
\title{transform cnv table to discrete form 0, 2}
\usage{
discretize_cnv(cnv_table)
}
\arguments{
\item{cnv_table}{matrix table of cnv}
}
\value{
discrete cnv table
}
\description{
transform cnv table to discrete form 0, 2
}
\examples{
 
}
